Greplin Challenge
Solution to Greplin challenge in Objective C.
There are 3 questions for this challenge:
- Question 1: Find the longest palendrome of the input string
- Question 2: Find the first prime fibonacci number greater than given
value and then find its prime factors
- Question 3: Find the number of subsets in a given array where sum of
al the elements is equalt to highest value in subset

=======
Solution has category extension for NSString, NSNumber and NSArray to
calculate the required values.
There is no UI as such, once executed the answers are printed in the
console.